The SetHeight and SetWeight methods were using int while the properties in Person use double.

+++ LTU-Lexicon24_övning3-OOP/PersonHandler.cs
@@ -11,12 +11,12 @@
             pers.Age = age;
         }
 
-        public void SetHeight(Person pers, int height) 
+        public void SetHeight(Person pers, double height) 
         {
             pers.Height = height;
         }
 
-        public void SetWeight(Person pers, int weight)
+        public void SetWeight(Person pers, double weight)
         {
             pers.Weight = weight;
         }
